feat: index completo + seed prisma com 20 produtos
