+++ .env
@@ -7,9 +7,11 @@ PRODUCT_SERVICE_PORT=3015
 
 NODE_ENV=development
 
-POSTGRES_DB="erp-product-coupon"
-POSTGRES_USER="user"
-POSTGRES_PASSWORD="password"
-POSTGRES_HOST="localhost"
+POSTGRES_DB=erp-product-coupon
+POSTGRES_USER=user
+POSTGRES_PASSWORD=password
+POSTGRES_HOST=localhost
 
-DATABASE_URL="postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}:5432/${POSTGRES_DB}?schema=public"
+DATABASE_URL="postgresql://user:password@localhost:5432/erp-product-coupon?schema=public"
+
+#DATABASE_URL="postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}:5432/${POSTGRES_DB}?schema=public"
