+++ README.md
@@ -2,13 +2,21 @@
 
 ### Como rodar
 
-A documentação pode ser vista ao acessar: http://localhost:3000/api/v1/docs
+A documentação pode ser vista ao acessar: http://localhost:3000/docs
 npm install -g nx
 pnpm install
 docker compose up -d
 pnpm prisma generate --schema=shared/prisma-config/prisma/schema.prisma
 pnpm run erp:migrate
+pnpm run erp:seed (caso o banco esteja vazio apenas)
 nx run-many --target=serve --projects=product-service,api-gateway (ou nx run apps/<nome do serviço>:serve)
+
+para resetar (caso ache necessário, OBS: vai apagar TODOS os containeres até mesmo os inativos de outros projetos):
+docker stop $(docker ps -aq) &&
+docker rm $(docker ps -aq) &&
+docker-compose down -v &&
+docker volume prune -f &&
+docker-compose up --build
 
 ### Fase de Pesquisa
