+++ shared/prisma-config/tsconfig.lib.json
@@ -11,5 +11,5 @@
     "forceConsistentCasingInFileNames": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src/**/*.ts"]
+  "include": ["src/**/*.ts", "prisma/seed.ts"]
 }
